requirements.txt
Django~=3.1
Pillow~=8.1.0
djangorestframework~=3.12.2
psycopg2-binary~=2.8.6
drf-yasg~=1.20.0
celery~=5.0.0
redis~=3.5.3
django-redis~=4.12.1
django-celery-beat~=2.2.0
coverage~=5.5
python-dotenv~=0.17.0
gunicorn==20.0.4
whitenoise==5.2.0